Split off macho relocs code
