+++ librz/bin/format/mach0/mach064_relocs.c
@@ -1,0 +1,5 @@
+// SPDX-FileCopyrightText: 2021 Florian Märkl <info@florianmaerkl.de>
+// SPDX-License-Identifier: LGPL-3.0-only
+
+#define RZ_BIN_MACH064 1
+#include "mach0_relocs.c"
