+++ librz/bin/meson.build
@@ -110,8 +110,10 @@ rz_bin_sources = [
   'format/mach0/coresymbolication.c',
   'format/mach0/fatmach0.c',
   'format/mach0/mach0.c',
+  'format/mach0/mach0_relocs.c',
   'format/mach0/mach0_rebase.c',
   'format/mach0/mach064.c',
+  'format/mach0/mach064_relocs.c',
   'format/mach0/mach064_rebase.c',
   'format/xnu/rz_cf_dict.c',
   'format/mdmp/mdmp.c',
