+++ test/db/formats/mach0/leb128
@@ -3,7 +3,7 @@ FILE=bins/mach0/macho-trie-bad-export-info-malformed-uleb128
 CMDS=<<EOF
 EOF
 EXPECT_ERR=<<EOF
-malformed uleb128
-malformed uleb128
+ERROR: malformed uleb128
+ERROR: malformed uleb128
 EOF
 RUN
